+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## Next release
 
 ### Added
+- The `-j/--jobs` flag for specifying number of subprocesses to use to run checks in parallel.
 - expression statements will now match by default also return statements
   ```
   foo();
@@ -13,8 +14,6 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
   ```javascript
   return foo();
   ```
-
-### Added
 - You can now use regexps for field names:
   ```
   var $X = {"=~/[lL]ocation/": $Y};